Make sorting the individual sections easier.
Making the individual entries each one line allows a section to be
sorted a lot more easily - we can now just use the editor's sort
function instead of doing it by hand.

Change the utilities entries to be "foo by bar" instead of "bar's foo"
so that the utility names are more easily sorted alphabetically.