Update dependencies and Travis CI config
This patch removes the explicit listing of jekyll-sass-converter
(included with Jekyll) and redcarpet (GitHub Pages doesn't support it
anyways and uses kramdown instead) and updates dependences to the latest
versions.

It also improves the Travis CI config by having it first build the site
and then check the generated site's links, rather than checking links in
the source file `index.md`. This doesn't make that much of a difference
now, though it does add checking of all the links in
`_includes/default.html`, which is nice. Once this site gets multiple
pages, this change will be especially useful.

This patch also updates the Google Analytics script to the latest one
provided from Google. This is not exactly necessary, but it has the nice
side effect of working around awesome_bot accidentally picking up the
'https://ssl' and 'http://www' from the old script as links.