Switch to GitHub Actions and proof-html
This patch adds a GitHub Action to check the HTML and links on every
push and PR, as well as on a weekly schedule. It also switches from
awesome_bot to proof-html (which uses
https://github.com/gjtorikian/html-proofer), which does a better job of
finding links by actually parsing the HTML rather than effectively
grepping for things that look like links.